dependencies: fix missing email_validator [more]
This is a 'quick' fix for the 2.3.1 breaking change in wtforms
that requires the email-validator library. WTForms is brought in
from flask-security and flask-security is a dependency of
invenio-accounts.

The real fix would be switching to 'flask-security-too` since
flask-security is not maintained anymore. However, a cursory attempt
has generated more errors. So until we can address those errors,
this fix can do.

- also fixes pytest dependency issues